Update Makefile (+3 more changes)

+++ build.sh
@@ -8,7 +8,7 @@ set -eou pipefail
 ZMK_IMAGE="zmkfirmware/zmk-build-arm:2.5-branch"
 
 main() {
-  local board='' shield_left='' shield_right='' GID
+  local board='' shield_left='' shield_right='' script='' GID
   GID="$(id -g)"
 
   while true; do
@@ -27,31 +27,30 @@ main() {
 
   rm -rf build/*.out
 
-  docker run -it --rm \
-    -v "$(pwd)/.cache:/keeb" \
-    -v "$(pwd)/config:/keeb/config:ro" \
-    -v "$(pwd)/build:/build" \
-    "$ZMK_IMAGE" sh -c "
-      cd /keeb
+  script="
+    cd /keeb
 
-      if [ ! -d .west ]; then
-        west init -l config
-      fi
+    if [ ! -d .west ]; then
+      west init -l config
+    fi
 
-      if [ ! -d zmk ]; then
-        west update
-      fi
+    if [ ! -d zmk ]; then
+      west update
+    fi
 
-      west zephyr-export
+    west zephyr-export
 
-      echo 'Building left'
-      west build --pristine -b $board zmk/app -- \
-        -DSHIELD=$shield_left \
-        -DZMK_CONFIG=/keeb/config
-      cp /keeb/build/zephyr/zmk.uf2 /build/${shield_left}_${board}.uf2.out
-      chown $UID:$GID /build/${shield_left}_${board}.uf2.out
-      chown $UID:$GID /build
+    echo 'Building left'
+    west build --pristine -b $board zmk/app -- \
+      -DSHIELD=$shield_left \
+      -DZMK_CONFIG=/keeb/config
+    cp /keeb/build/zephyr/zmk.uf2 /build/${shield_left}_${board}.uf2.out
+    chown $UID:$GID /build/${shield_left}_${board}.uf2.out
+    chown $UID:$GID /build
+  "
 
+  if [[ -n "$shield_right" ]]; then
+    script="$script
       echo 'Building right'
       west build --pristine -b $board zmk/app -- \
         -DSHIELD=$shield_right \
@@ -59,9 +58,15 @@ main() {
       cp /keeb/build/zephyr/zmk.uf2 /build/${shield_right}_${board}.uf2.out
       chown $UID:$GID /build/${shield_right}_${board}.uf2.out
     "
+  fi
 
-  if [[ ! -e "build/${shield_left}_${board}.uf2.out" ]] \
-    || [[ ! -e "build/${shield_right}_${board}.uf2.out" ]]; then
+  docker run -it --rm \
+    -v "$(pwd)/.cache:/keeb" \
+    -v "$(pwd)/config:/keeb/config:ro" \
+    -v "$(pwd)/build:/build" \
+    "$ZMK_IMAGE" sh -c "$script"
+
+  if [[ ! -e "build/${shield_left}_${board}.uf2.out" ]]; then
     echo ''
     echo '-- Errors in building'
     exit 1
@@ -69,11 +74,16 @@ main() {
 
   echo ''
   echo '-- Done'
-  rm -f build/${shield_left}_${board}.uf2 build/${shield_right}_${board}.uf2
+
+  rm -f build/${shield_left}_${board}.uf2
   mv build/${shield_left}_${board}.uf2{.out,}
-  mv build/${shield_right}_${board}.uf2{.out,}
   ls -la build/${shield_left}_${board}.uf2
-  ls -la build/${shield_right}_${board}.uf2
+
+  if [[ -f "build/${shield_right}_${board}.uf2.out" ]]; then
+    rm -f build/${shield_right}_${board}.uf2
+    mv build/${shield_right}_${board}.uf2{.out,}
+    ls -la build/${shield_right}_${board}.uf2
+  fi
 }
 
 main "$@"
